+++ qtjsonschema/widgets.py
@@ -2,10 +2,17 @@
 Widget definitions for JSON schema elements.
 """
 
+from abc import abstractmethod
+
 from PyQt5 import QtCore, QtWidgets, QtGui
 
-from abc import abstractmethod, ABC
+from .tools import CachedURILoaderRegistry, FileResourceLoader, HTTPResourceLoader, Context, DocumentLoader
 
+
+# Widgets supporting $ref
+# $ref, items, allOf, anyOf, additionalItems, dependencies,
+# oneOf, type, extends, properties, patternProperties,
+# additionalProperties
 
 def iter_widgets(layout):
     for i in range(layout.count()):
@@ -36,10 +43,6 @@ class UnsupportedSchema(QtWidgets.QLabel):
         return "(unsupported)"
 
 
-# $ref, items, allOf, anyOf, additionalItems, dependencies,
-# oneOf, type, extends, properties, patternProperties,
-# additionalProperties
-
 class JSONBaseWidget:
     def __init__(self, name, schema, ctx, parent):
         super().__init__()
@@ -56,13 +59,11 @@ class JSONBaseWidget:
         else:
             self.definitions = {}
 
-    @abstractmethod
     def from_json_object(self, data):
-        pass
+        raise NotImplementedError
 
-    @abstractmethod
     def to_json_object(self):
-        pass
+        raise NotImplementedError
 
 
 class JSONObjectWidget(JSONBaseWidget, QtWidgets.QGroupBox):
@@ -297,7 +298,7 @@ class JSONArrayWidget(JSONBaseWidget, QtWidgets.QWidget):
         schema = self._get_item_schema(index)
         obj = _create_widget("Item #{:d}".format(index), schema, self.ctx, self)
 
-        self.layout.addWidget(obj)
+        self.items_layout.addWidget(obj)
 
         if data is not None:
             obj.from_json_object(data)
@@ -310,7 +311,7 @@ class JSONArrayWidget(JSONBaseWidget, QtWidgets.QWidget):
                 self.add_item(datum)
 
     def to_json_object(self):
-        return [w.to_json_object() for w in iter_widgets(self.items_layout) if hasattr(w, "to_json_object")]
+        return [w.to_json_object() for w in iter_widgets(self.items_layout)]
 
 
 schema_type_to_widget_class = {
@@ -321,104 +322,6 @@ schema_type_to_widget_class = {
     "number": JSONNumberWidget,
     "boolean": JSONBooleanWidget
 }
-
-from abc import ABC, abstractmethod
-from functools import lru_cache
-from uritools import uricompose, urijoin, urisplit
-import requests
-from json import load as load_json
-
-
-class ResourceLoader(ABC):
-    @abstractmethod
-    def load_resource(self, location):
-        pass
-
-
-class HTTPResourceLoader(ResourceLoader):
-
-    def load_resource(self, location):
-        return requests.get(location).json()
-
-
-class FileResourceLoader(ResourceLoader):
-
-    def load_resource(self, location):
-        result = urisplit(location)
-        if result.authority:
-            raise ValueError("Network paths unsupported")
-
-        path = result.path[1:]
-        with open(path) as f:
-            return load_json(f)
-
-
-class DocumentLoader(ResourceLoader):
-
-    def __init__(self, document, location):
-        self.location = location
-        self.document = document
-
-    def load_resource(self, location):
-        if location != self.location:
-            raise ValueError("Cannot retrieve external documents")
-        return self.document
-
-
-class URILoaderRegistry:
-    def __init__(self):
-        self.scheme_to_loader = {}
-
-    def load_resource_from_loader(self, loader, location):
-        print(f"Loading resource {location} with {loader}")
-        return loader.load_resource(location)
-
-    def load_uri(self, uri):
-        result = urisplit(uri)
-        location = uricompose(result.scheme, result.authority, result.path)
-        loader = self.scheme_to_loader[result.scheme]
-        resource = self.load_resource_from_loader(loader, location)
-
-        if result.fragment:
-            assert result.fragment.startswith("/")
-            reference = Reference(result.fragment[1:])
-            return reference.extract(resource)
-
-        return resource
-
-    def register_for_scheme(self, scheme, resource):
-        self.scheme_to_loader[scheme] = resource
-
-
-class CachedURILoaderRegistry(URILoaderRegistry):
-    load_resource_from_loader = lru_cache(1024)(URILoaderRegistry.load_resource_from_loader)
-
-
-class Reference:
-    def __init__(self, uri):
-        self.elements = [e.replace('~1', '/').replace('~0', '~') for e in uri.split('/')]
-
-    def extract(self, obj):
-        for elem in self.elements:
-            obj = obj[elem]
-        return obj
-
-
-class Context:
-    def __init__(self, scope_uri, registry):
-        self.scope_uri = scope_uri
-        self.registry = registry
-
-    def follow_uri(self, uri):
-        new_uri = urijoin(self.scope_uri, uri)
-        return self.__class__(new_uri, self.registry)
-
-    def dereference(self, uri):
-        reference_path = urijoin(self.scope_uri, uri)
-        return self.registry.load_uri(reference_path)
-
-    def __repr__(self):
-        return f"Context({self.scope_uri!r}, {self.registry!r})"
 
 
 def create_widget(name, schema, schema_uri=None):
@@ -460,23 +363,3 @@ def _create_widget(name, schema, ctx, parent):
                 return UnsupportedSchema(name, schema, parent)
 
     return UnsupportedSchema(name, schema, parent)
-
-
-if __name__ == "__main__":
-    ref = f"identifier.json#/definitions/identifier"
-    registry = URILoaderRegistry()
-
-    http_retriever = HTTPResourceLoader()
-    registry.register_for_scheme('http', http_retriever)
-    registry.register_for_scheme('https', http_retriever)
-    registry.register_for_scheme('file', FileResourceLoader())
-
-    from pathlib import Path
-    uri = (Path(__file__).parent / 'schema.json').as_uri()
-    ctx = Context(uri, registry)
-    # ctx.follow_uri(schema_url)
-    result = ctx.dereference(ref)
-
-
-    # TODO "ID" is used as a base URL, we don't do that explictly atm, so where a reference doesn't
-    # give a scheme it will fail to join properly
